Added wall sliding, and HUD
Added HUD which displays the no. of items collected by the player and is made by user interface node and also added wall sliding of player meaning when it is not touching the floor but on side of tiles, it'll slow down and make it appear as if player sliding on the walls.